[add]添加功能
1. 支持弹幕过滤
2. 支持离线播放
3. 其它小优化

+++ Mac/AniXPlayer/Base/View/TableViewCell/SliderTableViewCell.swift
@@ -55,7 +55,7 @@ class SliderTableViewCell: NSView {
     @IBOutlet weak var valueSlider: NSSlider!
     
     /// 步长
-    var step: UInt = 0
+    var step: Float = 0
     
     var onChangeSliderCallBack: ((SliderTableViewCell) -> Void)?
